requirements.txt
Flask
python-dotenv
groq
gunicorn
PyPDF2
langchain
faiss-cpu
sentence-transformers
deep-translator